requirements.txt
pyinstaller
tkinter
Pillow
pylint
